[PyTorch] Establish codegen framework
Copied a subset of codegen scripts. We are able to generate lightweight dispatch code, but it is not tweaked to fit the EValue and native function system. Will add those in later PRs.

To run it, use the following command:
```
python codegen/gen.py --static_dispatch_backend=CPU --skip_dispatcher_op_registration --per-operator-headers --backend_whitelist=CPU
```

It will generate the following code:
* `RegisterCPU.cpp`, contains wrapper to native function.
* `UnboxingFunctionsEverything.cpp`, contains unboxing logic that connects boxed kernel to unboxed kernel in `RegisterCPU.cpp`.
* `RegisterCodegenUnboxedKernelsEverything.cpp`, contains C++ code that register all boxed kernels in `Unboxed...` mentioned above into op registry.

All the other generated code are useless and can be removed.

[ghstack-poisoned]